Restructure move kubernetes package to infrastructure/kubernetes
