+++ src/main/groovy/com/cloudogu/gitops/FeatureWithImage.groovy
@@ -1,7 +1,7 @@
 package com.cloudogu.gitops
 
 import com.cloudogu.gitops.config.Config
-import com.cloudogu.gitops.kubernetes.api.K8sClient
+import com.cloudogu.gitops.infrastructure.kubernetes.api.K8sClient
 
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
