+++ src/main/groovy/com/cloudogu/gitops/infrastructure/deployment/HelmStrategy.groovy
@@ -1,7 +1,7 @@
 package com.cloudogu.gitops.infrastructure.deployment
 
 import com.cloudogu.gitops.config.Config
-import com.cloudogu.gitops.kubernetes.api.HelmClient
+import com.cloudogu.gitops.infrastructure.helm.HelmClient
 
 import java.nio.file.Path
 import jakarta.inject.Singleton
