+++ src/main/groovy/com/cloudogu/gitops/infrastructure/git/providers/scmmanager/ScmManager.groovy
@@ -10,7 +10,7 @@ import com.cloudogu.gitops.infrastructure.git.providers.RepoUrlScope
 import com.cloudogu.gitops.infrastructure.git.providers.Scope
 import com.cloudogu.gitops.infrastructure.git.providers.scmmanager.api.Repository
 import com.cloudogu.gitops.infrastructure.git.providers.scmmanager.api.ScmManagerApiClient
-import com.cloudogu.gitops.kubernetes.api.K8sClient
+import com.cloudogu.gitops.infrastructure.kubernetes.api.K8sClient
 import com.cloudogu.gitops.utils.NetworkingUtils
 
 import groovy.util.logging.Slf4j
