+++ src/main/groovy/com/cloudogu/gitops/infrastructure/kubernetes/rbac/RbacDefinition.groovy
@@ -1,4 +1,4 @@
-package com.cloudogu.gitops.kubernetes.rbac
+package com.cloudogu.gitops.infrastructure.kubernetes.rbac
 
 import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.infrastructure.git.GitRepo
