+++ src/main/groovy/com/cloudogu/gitops/utils/NetworkingUtils.groovy
@@ -1,6 +1,6 @@
 package com.cloudogu.gitops.utils
 
-import com.cloudogu.gitops.kubernetes.api.K8sClient
+import com.cloudogu.gitops.infrastructure.kubernetes.api.K8sClient
 
 import jakarta.inject.Singleton
 import groovy.util.logging.Slf4j
