+++ src/test/groovy/com/cloudogu/gitops/application/content/ContentLoaderTest.groovy
@@ -17,7 +17,7 @@ import com.cloudogu.gitops.config.scm.ScmTenantSchema
 import com.cloudogu.gitops.features.Jenkins
 import com.cloudogu.gitops.infrastructure.deployment.DeploymentStrategy
 import com.cloudogu.gitops.infrastructure.git.GitRepoFactory
-import com.cloudogu.gitops.kubernetes.api.K8sClient
+import com.cloudogu.gitops.infrastructure.kubernetes.api.K8sClient
 import com.cloudogu.gitops.testhelper.git.GitHandlerForTests
 import com.cloudogu.gitops.testhelper.git.ScmManagerMock
 import com.cloudogu.gitops.testhelper.git.TestGitRepoFactory
