+++ src/test/groovy/com/cloudogu/gitops/cli/GitopsPlaygroundCliTest.groovy
@@ -9,7 +9,7 @@ import com.cloudogu.gitops.Application
 import com.cloudogu.gitops.config.ApplicationConfigurator
 import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.destroy.Destroyer
-import com.cloudogu.gitops.kubernetes.api.K8sClient
+import com.cloudogu.gitops.infrastructure.kubernetes.api.K8sClient
 
 import io.micronaut.context.ApplicationContext
