+++ src/test/groovy/com/cloudogu/gitops/features/JenkinsTest.groovy
@@ -9,11 +9,11 @@ import com.cloudogu.gitops.application.orchestration.GitHandler
 import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.config.scm.ScmTenantSchema
 import com.cloudogu.gitops.infrastructure.deployment.HelmStrategy
-import com.cloudogu.gitops.jenkins.GlobalPropertyManager
-import com.cloudogu.gitops.jenkins.JobManager
-import com.cloudogu.gitops.jenkins.PrometheusConfigurator
-import com.cloudogu.gitops.jenkins.UserManager
-import com.cloudogu.gitops.kubernetes.api.K8sClient
+import com.cloudogu.gitops.infrastructure.jenkins.GlobalPropertyManager
+import com.cloudogu.gitops.infrastructure.jenkins.JobManager
+import com.cloudogu.gitops.infrastructure.jenkins.PrometheusConfigurator
+import com.cloudogu.gitops.infrastructure.jenkins.UserManager
+import com.cloudogu.gitops.infrastructure.kubernetes.api.K8sClient
 import com.cloudogu.gitops.testhelper.git.GitHandlerForTests
 import com.cloudogu.gitops.testhelper.git.ScmManagerMock
 import com.cloudogu.gitops.utils.CommandExecutorForTest
