+++ src/test/groovy/com/cloudogu/gitops/infrastructure/kubernetes/api/K8sJavaApiTest.groovy
@@ -1,4 +1,4 @@
-package com.cloudogu.gitops.kubernetes.api
+package com.cloudogu.gitops.infrastructure.kubernetes.api
 
 import com.cloudogu.gitops.config.Credentials
