+++ src/test/groovy/com/cloudogu/gitops/utils/K8sClientTest.groovy
@@ -4,7 +4,7 @@ import static groovy.test.GroovyAssert.shouldFail
 import static org.assertj.core.api.Assertions.assertThat
 
 import com.cloudogu.gitops.config.Config
-import com.cloudogu.gitops.kubernetes.api.K8sClient
+import com.cloudogu.gitops.infrastructure.kubernetes.api.K8sClient
 
 import groovy.yaml.YamlSlurper
